+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Reduce memory allocation in case of initialization from big.Int [#252](https://github.com/shopspring/decimal/pull/252)
 
 #### BUGFIXES
-- Fix binary marshalling of decimal zero value  [#253](https://github.com/shopspring/decimal/pull/253)
+- Fix binary marshaling of decimal zero value  [#253](https://github.com/shopspring/decimal/pull/253)
 
 ## Decimal v1.3.0